cli: add cmd to add subdomain
